implementing cookies

+++ src/app/api/profile/[id]/route.ts
@@ -2,6 +2,7 @@ import prisma from "@/utils/db";
 import { NextRequest, NextResponse } from "next/server";
 import jwt from "jsonwebtoken";
 import { JWTPayload } from "@/utils/types";
+import { verifyToken } from "@/utils/verifyToken";
 
 interface Props {
   params: { id: string };
@@ -21,10 +22,9 @@ export async function DELETE(request: NextRequest, { params }: Props) {
         if (!user) {
             return NextResponse.json({ message: "User Not Found" }, { status: 404 });
         }
-
-        const authToken = request.headers.get('authToken') as string
-        const userFromToken = jwt.verify(authToken, process.env.JWT_TOKEN as string) as JWTPayload
-        if (userFromToken.id === user.id) {
+        
+        const userFromToken = verifyToken(request)
+        if (userFromToken !== null && userFromToken.id === user.id) {
             await prisma.user.delete({ where: { id: parseInt(params.id) } });
             return NextResponse.json(
                 { message: "Your Profile Has Been Deleted" },
